feat: cleanup Acknowledgement.tsx (#8472)
