+++ src/components/Acknowledgement/InfoAcknowledgement.tsx
@@ -1,0 +1,5 @@
+import type { AcknowledgementProps } from './Acknowledgement'
+import { Acknowledgement } from './Acknowledgement'
+
+export const InfoAcknowledgement = (props: AcknowledgementProps) =>
+  Acknowledgement({ ...props, buttonColorScheme: 'blue', iconColorScheme: 'yellow' })
