+++ src/components/TradeAssetSearch/components/CustomAssetAcknowledgement.tsx
@@ -15,7 +15,7 @@ import type { Asset } from '@shapeshiftoss/types'
 import type * as CSS from 'csstype'
 import { useCallback, useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
-import { WarningAcknowledgement } from 'components/Acknowledgement/Acknowledgement'
+import { WarningAcknowledgement } from 'components/Acknowledgement/WarningAcknowledgement'
 import { AssetIcon } from 'components/AssetIcon'
 import { InlineCopyButton } from 'components/InlineCopyButton'
 import { useToggle } from 'hooks/useToggle/useToggle'
